requirements.txt
aiodns==1.1.1
aiofiles==0.3.0
aiohttp==1.0.5
aiohttp-jinja2==0.8.0
aiopg==0.12.0
async-timeout==1.1.0
Babel==2.3.4
cchardet==1.1.0
chardet==2.3.0
Jinja2==2.8
MarkupSafe==0.23
multidict==2.1.2
pkg-resources==0.0.0
psycopg2==2.6.2
pycares==2.1.1
pycrypto==2.6.1
pytz==2016.7
PyYAML==3.12
SQLAlchemy==1.1.3
